perf(canvas): 画布平移 LOD 外壳与 VideoNode 面板拆分
根因:大画布平移帧率受制于移动 transform 子树的渲染树规模(302 节点
≈13.5k 元素,每帧主线程遍历 ~41ms),与绘制内容无关——隐藏全部图片/
边线/hover 均无改善,内容子树摘除即回满帧。

- zoom<0.35 节点渲染 4 元素轻外壳(withLodShell):显式尺寸+同位同 id
  的真实 Handle(边锚定不变,规避 content-visibility 挂载期垃圾
  handleBounds 永不自愈的问题)+缩略图+生成中角标;豁免 skill/group/
  beatContext 与播放中的视频
- 低缩放档关闭 onlyRenderVisibleElements(挂载翻腾 843 次/4s→0);
  升档与 shell→full 交换同帧提交,降档推迟到 moveEnd
- shell→完整组件经升级队列每帧放行 3 个,手势中不放行:快速平移的
  挂载尖刺 >25ms 帧 6→1/200,升档最长帧 964→146ms
- 手势期/低缩放档 CSS 降级(backdrop-filter/shadow/filter/边动画),
  节点标题溢出测量延迟到画布静止
- 离屏 CORS 抓帧管道合一(videoFrameCapture):用户抓帧与 LOD 视频
  缩略图共用,生产跨域可用,并发 2/空闲调度/15s 超时
- VideoNode 生成面板拆出 VideoOperationsPanel(选中才挂载):主体
  5309→3480 行,挂载路径少 ~25 hook,估价请求仅选中时发起;提交链/
  自动纠正 effect/mention 重编号等留主体保证未选中行为不变

实测(302 节点/zoom 0.1 快速平移):17.2→71.9fps(75Hz 顶满),
p90 128.6→15.4ms,p99 470→21.3ms。
