+++ frontend/src/features/canvas/nodes/index.ts
@@ -2,6 +2,7 @@
 // Copyright (c) 2026 ClaymoreLab
 import type { NodeTypes } from '@xyflow/react';
 
+import { withLodShell } from './LodShellNode';
 import { AudioNode } from './AudioNode';
 import { BeatContextNode } from './BeatContextNode';
 import { GroupNode } from './GroupNode';
@@ -20,24 +21,26 @@ import { VideoComposeNode } from './VideoComposeNode';
 import { VideoNode } from './VideoNode';
 import { VideoStoryNode } from './VideoStoryNode';
 
+// 全部经 withLodShell 包装：低缩放档渲染轻量外壳（豁免类型在包装器内部判断，
+// 保持这张表均质）。包装发生在模块加载期，引用稳定，不会造成节点重挂。
 export const nodeTypes: NodeTypes = {
-  audioNode: AudioNode,
-  beatContextNode: BeatContextNode,
-  exportImageNode: ImageNode,
-  groupNode: GroupNode,
-  imageGenNode: ImageGenNode,
-  imageNode: ImageEditNode,
-  pano360ViewerNode: Pano360ViewerNode,
-  scriptNode: ScriptNode,
-  skillNode: SkillNode,
-  storyboardGenNode: StoryboardGenNode,
-  storyboardNode: StoryboardNode,
-  textAnnotationNode: TextAnnotationNode,
-  threeDWorldNode: ThreeDWorldNode,
-  uploadNode: UploadNode,
-  videoComposeNode: VideoComposeNode,
-  videoNode: VideoNode,
-  videoStoryNode: VideoStoryNode,
+  audioNode: withLodShell('audioNode', AudioNode),
+  beatContextNode: withLodShell('beatContextNode', BeatContextNode),
+  exportImageNode: withLodShell('exportImageNode', ImageNode),
+  groupNode: withLodShell('groupNode', GroupNode),
+  imageGenNode: withLodShell('imageGenNode', ImageGenNode),
+  imageNode: withLodShell('imageNode', ImageEditNode),
+  pano360ViewerNode: withLodShell('pano360ViewerNode', Pano360ViewerNode),
+  scriptNode: withLodShell('scriptNode', ScriptNode),
+  skillNode: withLodShell('skillNode', SkillNode),
+  storyboardGenNode: withLodShell('storyboardGenNode', StoryboardGenNode),
+  storyboardNode: withLodShell('storyboardNode', StoryboardNode),
+  textAnnotationNode: withLodShell('textAnnotationNode', TextAnnotationNode),
+  threeDWorldNode: withLodShell('threeDWorldNode', ThreeDWorldNode),
+  uploadNode: withLodShell('uploadNode', UploadNode),
+  videoComposeNode: withLodShell('videoComposeNode', VideoComposeNode),
+  videoNode: withLodShell('videoNode', VideoNode),
+  videoStoryNode: withLodShell('videoStoryNode', VideoStoryNode),
 };
 
 export { AudioNode, BeatContextNode, GroupNode, ImageEditNode, ImageGenNode, ImageNode, Pano360ViewerNode, ScriptNode, SkillNode, StoryboardGenNode, StoryboardNode, TextAnnotationNode, ThreeDWorldNode, UploadNode, VideoComposeNode, VideoNode, VideoStoryNode };
