+++ frontend/src/features/canvas/ui/NodeHeader.tsx
@@ -9,6 +9,10 @@ import {
   useRef,
   useState,
 } from 'react';
+import {
+  isCanvasMeasurementDeferred,
+  onCanvasMeasurementResume,
+} from '@/features/canvas/application/canvasLod';
 
 type HeaderAdjust = {
   x?: number;
@@ -125,7 +129,14 @@ export function NodeHeader({
       return;
     }
 
+    // 读 scrollWidth/clientWidth 会强制同步布局，在画布这个体量的 DOM 上单次
+    // 约 2ms。快速平移时可见性剔除每帧要挂载 3~4 个节点，每个都跑一遍，直接
+    // 把帧预算吃穿（实测拖拽 20fps）。所以画布在动、或缩得太小看不清字的时候
+    // 一律不量，等它静下来再补一次——标题淡出效果本来也只有静止时才看得见。
     const measureOverflow = () => {
+      if (isCanvasMeasurementDeferred()) {
+        return;
+      }
       const nextOverflowing =
         titleMeasureElement.scrollWidth - titleMeasureElement.clientWidth > 1;
       setIsTitleOverflowing((previous) =>
@@ -137,10 +148,12 @@ export function NodeHeader({
     const frameId = requestAnimationFrame(measureOverflow);
     const observer = new ResizeObserver(measureOverflow);
     observer.observe(titleMeasureElement);
+    const unsubscribeResume = onCanvasMeasurementResume(measureOverflow);
 
     return () => {
       cancelAnimationFrame(frameId);
       observer.disconnect();
+      unsubscribeResume();
     };
   }, [isEditingTitle, titleMeasureElement, titleText]);
